+++ benchmarks/plasma-v0.0.x/test.js
@@ -7,11 +7,12 @@ plasma.on('do work', function (c, next) {
   next(null, c)
 })
 var elements = []
-for (var i = 0; i<10000; i++)
+for (var i = 0; i < 10000; i++) {
   elements.push(i)
+}
 async.each(elements, function (e, done) {
   plasma.emit('do work', function (err, result) {
-    done()
+    done(err, result)
   })
 }, function () {
   console.timeEnd('v0.0.7')